basic flask frontend integration
