+++ data/artifacts/.gitignore
@@ -1,2 +1,3 @@
 movie_list.pkl
 similarity.pkl
+
