+++ static/style.css
@@ -1,0 +1,149 @@
+:root {
+	--clr-primary: #edf0fe;
+	--clr-secondary: #ff405c;
+	--clr-tertiary: #53b9c6;
+}
+
+html {
+	scroll-behavior: smooth;
+}
+
+body {
+	/* background-color: #08080c; */
+	background: linear-gradient(180deg, #101016 0%, #0F0F17 34%, #10101C 49%, #0a0f12 67%, #12121F 83%, #08080C 100%);
+	color: azure;
+	font-family: "Poppins";
+	font-size: 1rem;
+	padding-inline: 5rem;
+	padding-block: 2rem;
+}
+
+*,
+*::after,
+*::before {
+	box-sizing: border-box;
+	margin: 0;
+	padding: 0;
+}
+
+h1 {
+	font-size: 3rem;
+	font-weight: 800;
+	color: var(--clr-secondary);
+	margin-bottom: 1rem;
+}
+
+p {
+	font-size: 1rem;
+	font-weight: 400;
+	color: var(--clr-primary);
+}
+
+h2 {
+	font-size: 2rem;
+	font-weight: 600;
+	color: var(--clr-tertiary);
+	margin-bottom: 1rem;
+}
+
+.title {
+	display: flex;
+	flex-direction: column;
+}
+
+.menu-container {
+	width: 1.5rem;
+	height: 2rem;
+	right: 5%;
+	top: 10%;
+	position: absolute;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	gap: 0.3rem;
+	display: flex;
+}
+
+.menu {
+	width: 100%;
+	border: 2px #53b9c6 solid;
+}
+
+.section {
+	display: flex;
+	flex-direction: column;
+	justify-content: flex-start;
+	align-items: flex-start;
+	gap: 8px;
+}
+
+.section-title {
+	color: #edf0fe;
+	font-size: 2rem;
+	font-weight: 500;
+
+}
+
+.search-container {
+    display: flex;
+    background: #edf0fe42;
+    padding: 0.4rem;
+    padding-inline: 1rem;
+    border-radius: 24px;
+    margin-bottom: 1rem;
+}
+
+.search-bar {
+    background: transparent;
+    border: none;
+    color: #edf0fe;
+    font-size: 1rem;
+    font-family: Poppins;
+    font-weight: 500;
+
+    width: 100%;
+}
+
+input:focus {
+    outline: none;
+}
+
+.img-container {
+	display: inline-flex;
+	justify-content: flex-start;
+	align-items: flex-start;
+	gap: 36.11px;
+}
+
+.img-placeholder {
+	width: 244px;
+	height: 293px;
+	position: relative;
+	background: linear-gradient(0deg, white 0%, white 100%);
+	box-shadow: 0px 4px 8.100000381469727px 7px rgba(0, 0, 0, 0.25);
+	border-radius: 11.11px;
+	content: url("https://via.placeholder.com/244x293");
+}
+
+.footer {
+    display: inline-flex;
+	width: 100%;
+	height: 4rem;
+    padding-block: 23px;
+    margin-top: 2rem;
+	background: linear-gradient(
+		180deg,
+		#151516 0%,
+		rgba(20.61, 20.61, 22.09, 0) 100%
+	);
+	border-top-left-radius: 2rem;
+	border-top-right-radius: 2rem;
+	overflow: hidden;
+	justify-content: center;
+	align-items: center;
+}
+
+.footer-text {
+	color: #4f4f4f;
+
+}
